requirements-test.txt
pytest==6.2.5
pytest-django==4.5.2
factory-boy==3.2.1
pytest-factoryboy==2.5.1